Move out python related code from column store core
