+++ cpp/arcticdb/column_store/string_pool.cpp
@@ -156,12 +156,6 @@ const Buffer& StringPool::shapes() const {
 
 size_t StringPool::size() const { return block_.size(); }
 
-py::buffer_info StringPool::as_buffer_info() const {
-    return py::buffer_info{
-            (void*)block_.at(0).data(), 1, py::format_descriptor<char>::format(), ssize_t(block_.at(0).size())
-    };
-}
-
 std::optional<position_t> StringPool::get_offset_for_column(std::string_view string, const Column& column) const {
     auto unique_values = unique_values_for_string_column(column);
     remove_nones_and_nans(unique_values);
