+++ cpp/arcticdb/column_store/string_pool.hpp
@@ -17,12 +17,6 @@
 #include <string_view>
 #include <unordered_set>
 
-namespace pybind11 {
-struct buffer_info;
-}
-
-namespace py = pybind11;
-
 #include <ankerl/unordered_dense.h>
 
 namespace arcticdb {
@@ -167,8 +161,6 @@ class StringPool {
     size_t size() const;
 
     [[nodiscard]] size_t num_blocks() const;
-
-    py::buffer_info as_buffer_info() const;
 
     std::optional<position_t> get_offset_for_column(std::string_view str, const Column& column) const;
     ankerl::unordered_dense::set<position_t> get_offsets_for_column(
